+++ src/EventStore.ClientAPI/ClientOperations/SubscriptionOperation.cs
@@ -220,7 +220,7 @@ namespace EventStore.ClientAPI.ClientOperations
             }
         }
 
-        protected void ConfirmSubscription(long lastCommitPosition, int? lastEventNumber)
+        protected void ConfirmSubscription(long lastCommitPosition, long? lastEventNumber)
         {
             if (lastCommitPosition < -1)
                 throw new ArgumentOutOfRangeException("lastCommitPosition", string.Format("Invalid lastCommitPosition {0} on subscription confirmation.", lastCommitPosition));
@@ -234,7 +234,7 @@ namespace EventStore.ClientAPI.ClientOperations
             _source.SetResult(_subscription);
         }
 
-        protected abstract T CreateSubscriptionObject(long lastCommitPosition, int? lastEventNumber);
+        protected abstract T CreateSubscriptionObject(long lastCommitPosition, long? lastEventNumber);
 
         protected void EventAppeared(ResolvedEvent e)
         {
